# Conflicts:
#	backend/src/main/java/boco/service/rental/ListingService.java

+++ backend/src/main/java/boco/service/rental/ListingService.java
@@ -6,6 +6,8 @@ import boco.models.rental.*;
 import boco.repository.profile.ProfileRepository;
 import boco.repository.rental.CategoryTypeRepository;
 import boco.repository.rental.ImageRepository;
+import boco.repository.rental.ImageRepository;
+import boco.repository.rental.LeaseRepository;
 import boco.repository.rental.ListingRepository;
 import boco.service.security.JwtUtil;
 import org.slf4j.Logger;
@@ -27,6 +29,7 @@ public class ListingService {
     private final ListingRepository listingRepository;
     private final ProfileRepository profileRepository;
     private final CategoryTypeRepository categoryTypeRepository;
+    private final LeaseRepository leaseRepository;
     private final ImageRepository imageRepository;
 
     Logger logger = LoggerFactory.getLogger(ListingService.class);
@@ -35,11 +38,11 @@ public class ListingService {
     private JwtUtil jwtUtil;
 
     @Autowired
-    public ListingService(ListingRepository listingRepository, ProfileRepository profileRepository,
-                          CategoryTypeRepository categoryTypeRepository, ImageRepository imageRepository) {
+    public ListingService(ListingRepository listingRepository, ProfileRepository profileRepository, CategoryTypeRepository categoryTypeRepository, LeaseRepository leaseRepository, ImageRepository imageRepository) {
         this.listingRepository = listingRepository;
         this.profileRepository = profileRepository;
         this.categoryTypeRepository = categoryTypeRepository;
+        this.leaseRepository = leaseRepository;
         this.imageRepository = imageRepository;
     }
 
@@ -132,7 +135,6 @@ public class ListingService {
             Listing savedListing = listingRepository.save(newListing);
             return new ResponseEntity<>(new ListingResponse(savedListing), HttpStatus.CREATED);
         } catch (Exception e) {
-            System.out.println(e.getMessage());
             return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
@@ -199,13 +201,37 @@ public class ListingService {
 
 
             System.out.println("Your id: " + profile.get().getId() + ", database id: " + listingData.get().getProfile().getId());
+
+            Optional<Listing> emptyListing = listingRepository.findById(1L);
+            if (!emptyListing.isPresent()){
+                return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+            }
+            setListingWhenDeleted(listingId, emptyListing.get());
             //TODO fix
-            System.out.println("Updated stuff" + listingRepository.updateLeaseWhenListingDeleted(listingId));
             listingRepository.deleteById(listingId);
+            System.out.println("Updated stuff" );
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (Exception e) {
             return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
         }
+    }
+
+    public void setListingWhenDeleted(Long listingId, Listing emptyListing){
+        List<Lease> leases = leaseRepository.getLeasesByListing_Id(listingId);
+
+        for (Lease lease :
+                leases) {
+            lease.setListing(emptyListing);
+        }
+        leaseRepository.saveAll(leases);
+
+        List<Image> images = imageRepository.getImageByListing_Id(listingId);
+
+        for (Image image :
+                images) {
+            image.setListing(emptyListing);
+        }
+        imageRepository.saveAll(images);
     }
 
     public static List<ListingResponse> convertListings(List<Listing> listings){
